Make size check of value a default option rather than a requirement
Although this is a reasonable setting for use of the client with the standard memcached server, Tokyo Tyrant also is able to use the memcached protocol but doesn't have the 1MB size limit.  This commit makes the size check on data optional for alternate and valid uses of the memcache-client library in connecting to different key-value stores.  Tests included both for the new optional setting and for the previously-untested size check.

Signed-off-by: Mike Perham <mperham@gmail.com>