Enforce socket timeout on connect
Earlier changes to the memcache client changed the timeout mechanism
from using the (somewhat unreliable) Timeout or SystemTimer libraries to
using low-level non-blocking I/O. This works great - assuming the client
connects in the first place.

This change re-introduces the timeout mechanism and wraps the socket
creation - if you try to connect to a memcache server that's
down/rebooting/whatever, it will timeout sooner than the OS-specified
timeout for the socket.